Refactored the package of expenseType model
