+++ src/main/java/com/cwallet/CHAMPWallet/models/account/Wallet.java
@@ -1,6 +1,6 @@
 package com.cwallet.CHAMPWallet.models.account;
 
-import com.cwallet.CHAMPWallet.models.expenseType.ExpenseType;
+import com.cwallet.CHAMPWallet.models.expense.ExpenseType;
 import com.cwallet.CHAMPWallet.models.income.IncomeType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
