+++ src/main/java/com/cwallet/CHAMPWallet/repository/expenseType/ExpenseTypeRepository.java
@@ -1,9 +1,8 @@
 package com.cwallet.CHAMPWallet.repository.expenseType;
 
-import com.cwallet.CHAMPWallet.models.expenseType.ExpenseType;
+import com.cwallet.CHAMPWallet.models.expense.ExpenseType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface ExpenseTypeRepository extends JpaRepository<ExpenseType, Long> {
+
 }
