+++ src/main/java/com/cwallet/CHAMPWallet/service/expenseType/impl/ExpenseTypeServiceImpl.java
@@ -3,7 +3,7 @@ package com.cwallet.CHAMPWallet.service.expenseType.impl;
 import com.cwallet.CHAMPWallet.dto.expenseType.ExpenseTypeDto;
 import com.cwallet.CHAMPWallet.models.account.UserEntity;
 import com.cwallet.CHAMPWallet.models.account.Wallet;
-import com.cwallet.CHAMPWallet.models.expenseType.ExpenseType;
+import com.cwallet.CHAMPWallet.models.expense.ExpenseType;
 import com.cwallet.CHAMPWallet.repository.expenseType.ExpenseTypeRepository;
 import com.cwallet.CHAMPWallet.security.SecurityUtil;
 import com.cwallet.CHAMPWallet.service.expenseType.ExpenseTypeService;
@@ -13,10 +13,8 @@ import org.springframework.stereotype.Service;
 import static com.cwallet.CHAMPWallet.mappers.expenseType.ExpenseTypeMapper.mapToExpenseType;
 @Service
 public class ExpenseTypeServiceImpl implements ExpenseTypeService {
-
     private ExpenseTypeRepository expenseTypeRepository;
     private SecurityUtil securityUtil;
-
     @Autowired
     public ExpenseTypeServiceImpl(ExpenseTypeRepository expenseTypeRepository, SecurityUtil securityUtil){
         this.expenseTypeRepository = expenseTypeRepository;
@@ -31,8 +29,5 @@ public class ExpenseTypeServiceImpl implements ExpenseTypeService {
         expense.setWallet(newWallet);
         expense.setEnabled(true);
         expenseTypeRepository.save(expense);
-
-
-
     }
 }
